HU04 Change getUserByEmail method
